This is a fork for zetbios and zetup
There are two main changes in this fork:
1) a port of the bios to watcom, it compiles and builds but
hangs on booting from disk, still needs debugging.

2) Zetup is the Zet DE1 board setup utility. It is not finished,
I would like to ad SD card setup function but that is not done yet.
I also included a BMP to MIF converter utility that will let you
customize the welcome screen that gets displayed on rom the VGA
of the DE1.

More to come....