My bad, I copied a bad make on the last push
Appologies to anyone who got that last bad makefile. I copied a test
one over that is no good for doing a full make. So sorry. This one
should be good. It does a full make and is commented pretty well too
I think (Something I always appreciate in a makefile :))